+++ packages/system/src/components/text-input.tsx
@@ -1,12 +1,12 @@
 import classNames from "classnames"
 import React from "react"
 import { useId } from "@reach/auto-id"
-import { sprinkles } from "../sprinkles/sprinkles.css"
 import {
 	textInputBoxStyle,
 	textInputContainerStyle,
 	textInputLabelStyle,
 } from "./text-input.css"
+import { visuallyHidden } from "../styles/utilities.css"
 
 export interface TextInputProps
 	extends React.ComponentPropsWithoutRef<"input"> {
@@ -39,7 +39,7 @@ export const TextInput = React.forwardRef<HTMLInputElement, TextInputProps>(
 					htmlFor={id}
 					className={classNames(
 						textInputLabelStyle,
-						hideLabel && sprinkles({ hidden: "visually" })
+						hideLabel && visuallyHidden
 					)}
 				>
 					{label}