+++ packages/system/src/icons/burger-icon.tsx
@@ -1,0 +1,20 @@
+import classNames from "classnames"
+import React from "react"
+import { iconStyle } from "./icon.css"
+import { IconProps } from "./shared-icon-types"
+
+export const BurgerIcon = React.forwardRef<SVGSVGElement, IconProps>(
+	function BurgerIcon({ className, size, ...props }, ref) {
+		return (
+			<svg
+				className={classNames(className, iconStyle({ size }))}
+				viewBox="0 0 24 24"
+				fill="currentColor"
+				ref={ref}
+				{...props}
+			>
+				<path d="M3 18H21V16H3V18ZM3 13H21V11H3V13ZM3 6V8H21V6H3Z" />
+			</svg>
+		)
+	}
+)